Nested Arrays - For Loop
Using the nested array seatingChart, I wanted to separate each array and create a variable team 1, team 2, team 3.
In each team I have allocated the index of 0 to be the team captain.

My question is can I set a team captain without having to repeat the name again ?
Also  when I created the variable  team (1,2,3), does this become an array that stores the names of seatingChart[0][1][2]?
Let say I just wanted to access team 1 in the console so it displays team 1 instead of running through all the other teams again, will this possible? if not does this have something to do with the scope of the for loop?