DBZ setTimeout and setInterval Functions
The above code goes with the youtube video (DBZ scene over 9000!), uses the setTimeout function with that keeps time with the dialogue of the scene. 
What you need to do is to set the time of the video to just over 15 seconds and then start the code in the console. Ideally have the Chrome window next to the youtube video so you can follow. It will give you some time to get ready and then it will give you a count down from 3, as soon as it displays ‘starts in 1’ press play. 
It may take a few tries to get it. Code was tested on the Chrome console.